refactor(graphql/user): clean-up the structure with separating different purpose resolvers and typeDefs
