+++ src/graphql/typeDefs/common/scalars.typeDefs.ts
@@ -1,0 +1,7 @@
+import { gql } from "graphql-tag";
+
+const scalarsTypeDefs = gql`
+  scalar JSON
+`;
+
+export default scalarsTypeDefs;
